+++ selene/common/predicate.py
@@ -72,12 +72,27 @@ def includes_word(expected, ignore_case=False):
     return lambda actual: expected in re.split(r'\s+', actual) if not ignore_case else includes_ignoring_case(expected)
 
 
-seq_compare_by = lambda f: lambda x, *xs: lambda y, *ys: \
-    True if x is None and y is None else bool(f(x)(y)) and seq_compare_by(f)(*xs or (None, ))(*ys or (None, ))
+# will not work with empty seqs :( todo: fix
+# currently we use it only for non-empty seqs taking this into account
+seq_compare_by = lambda f: lambda x=None, *xs: lambda y=None, *ys: \
+    True if x is None and y is None else bool(f(x)(y)) and seq_compare_by(f)(*xs)(*ys)
+
+
+# def seq_compare_by_2(f):
+#     def fn(x, *xs):
+#         def fn(y, *ys):
+#             return True if x is None and y is None else \
+#                 bool(f(x)(y)) and seq_compare_by(f)(*xs or (None, ))(*ys or (None, ))
+#         return fn
+#     return fn
 
 
 list_compare_by = lambda f: lambda expected: lambda actual: \
     seq_compare_by(f)(*expected)(*actual)
+
+
+# list_compare_by = lambda f: lambda expected: lambda actual: \
+#     seq_compare_by(f)(* expected if expected else (None,))(* actual if actual else (None, ))
 
 
 equals_to_list = list_compare_by(equals)